correctly initialize luasocket (avoid SIGPIPE) and check that a unix socket hasn't already been created

+++ script/load.d/2200-serverexec.lua
@@ -10,9 +10,16 @@
 local Lr = require"utils.lambda".Lr
 
 local fname = require"std.servertag".fntag .. "serverexec"
-os.execute("rm -f " .. fname)
-local ok, unix, cmdpipe = pcall(Lr"require'socket.unix', require'std.cmdpipe'")
+local ok, _, unix, cmdpipe = pcall(Lr"require 'socket', require'socket.unix', require'std.cmdpipe'")
 if not ok then return engine.writelog("Cannot load luasocket with unix socket domain support, serverexec will not be available:\n" .. unix) end
+local test = unix()
+if test:connect(fname) then
+  test:close()
+  engine.writelog("There is already a unix socket named " .. fname .. ", will not overwrite.")
+  return
+end
+test:close()
+os.execute("rm -f " .. fname)
 local pipein = unix()
 if pipein:bind(fname) ~= 1 or pipein:listen() ~= 1 then
   pipein:close()
